Swapped in tosig.cpp definition.
