+++ src/ToSig.cpp
@@ -129,12 +129,20 @@ namespace {
 		fn0001(VECTOR& ans):_ans(ans)
 		{
 		}
-
+#ifndef LIBALGEBRA_VECTORS_H
 		template <class T>
-	void operator()(T& element)
-	{
-		_ans[KeyToIndex<TENSOR, WIDTH>(element.first)] = element.second;
-	}
+        void operator()(T& element)
+        {
+            _ans[KeyToIndex<TENSOR, WIDTH>(element.first)] = element.second;
+        }
+#else
+        template <typename T>
+        void operator()(T& element)
+        {
+		    _ans[KeyToIndex<TENSOR, WIDTH>(element.key())] = element.value();
+        }
+#endif
+
 	};
 
 	//[&ans] (const decltype(*(arg.begin()))& element){
